Getting ng-color-picker pagackage (requried some updates to ngrx version)
