+++ package.json
@@ -25,10 +25,11 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.32",
     "@fortawesome/free-regular-svg-icons": "^5.15.1",
     "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@ngrx/effects": "^10.0.1",
-    "@ngrx/store": "^10.0.1",
+    "@ngrx/effects": "^11.1.1",
+    "@ngrx/store": "^11.1.1",
     "firebase": "^8.1.1",
     "include-media": "^1.4.9",
+    "ngx-color-picker": "^11.0.0",
     "ngx-dropzone": "^2.3.0",
     "rxjs": "~6.6.3",
     "tslib": "^2.0.0",
@@ -39,12 +40,12 @@
     "@angular/cli": "^11.0.2",
     "@angular/compiler-cli": "~11.0.2",
     "@angular/language-service": "~11.0.2",
-    "@ngrx/store-devtools": "^10.0.1",
+    "@ngrx/store-devtools": "^11.1.1",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
+    "jasmine-core": "3.8",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~5.1.1",
     "karma-chrome-launcher": "~3.1.0",
