+++ src/app/root-store/reducer.ts
@@ -1,19 +1,33 @@
 import { createReducer, on, Action, ActionReducer } from '@ngrx/store';
 import { initialState, AuthenticationState, State } from './state';
 import * as Actions from './actions';
-import { CreateEmailUserAdjuster } from './state-adjusters/createEmailUserAdjuster';
-import { NewFrameAdjuster } from './state-adjusters/newFrameAdjuster';
 import { FrameStoreActions } from './frame-store';
+import { UserFrameMetadata } from '../shared/models/firebase-collections/user';
+import { FramePermissions } from '../shared/models/constants/framePermissions';
+import { FrameModel } from '../shared/models/view-models/frameModel';
 
 
 
-
+// TODO: SHOULD NOT DO DATA TRANSLATION HERE...EFFECT SHOULD DISPATCH MULTIPLE ACTIONS WITH CORRECT STATE DATA
+function addFrameToUser(state: AuthenticationState, response: FrameModel): AuthenticationState {
+  const newFrame: UserFrameMetadata = {
+    frameId: response.id,
+    permissions: [FramePermissions.creator],
+    name: response.name
+  };
+  return { ...state, currentUser:
+    { ...state.currentUser, frames: [...state.currentUser.frames, newFrame] }
+  };
+}
 
 const reduce = createReducer(
     initialState,
-    CreateEmailUserAdjuster.RequestHandler,
-    CreateEmailUserAdjuster.SuccessHandler,
-    CreateEmailUserAdjuster.FailureHandler,
+    on(Actions.CreateEmailUser.Request, (state: AuthenticationState) =>
+      ({ ...state, isLoading: true })),
+    on(Actions.CreateEmailUser.RequestSuccess, (state: AuthenticationState, { successResponse }) =>
+      ({ ...state, currentUser: successResponse, isLoading: false, registerErrorMessage: null })),
+    on(Actions.CreateEmailUser.RequestFailure, (state: AuthenticationState, { failureResponse }) =>
+      ({ ...state, currentUser: null, isLoading: false, registerErrorMessage: failureResponse })),
 
     on(Actions.SignInWithEmail.Request, (state: AuthenticationState) =>
       ({ ...state, isLoading: true })),
@@ -28,9 +42,7 @@ const reduce = createReducer(
       ({ ...state, currentUser: successResponse, isLoading: false, loginErrorMessage: null })),
     on(Actions.GetUserDataFromSignedInUser.RequestFailure, (state: AuthenticationState, { failureResponse }) =>
       ({ ...state, currentUser: null, isLoading: false, loginErrorMessage: failureResponse })),
-
-    NewFrameAdjuster.SuccessHandler,
-
+    on(FrameStoreActions.NewFrame.RequestSuccess, (state: AuthenticationState, { successResponse }) => addFrameToUser(state, successResponse)),
     on(FrameStoreActions.JoinFrame.RequestSuccess, (state: AuthenticationState, { successResponse }) =>
       ({ ...state, currentUser: { ...state.currentUser, frames: [...state.currentUser.frames, successResponse] }}))
 );
